Fix crash when deleting without items
If the UI thread blocks for a brief period of time and you press the
delete button multiple times, the callback will all be run sequentially.
So we need to ensure this function does not crash.

+++ pupil_src/shared_modules/gaze_producer/ui/storage_edit_menu.py
@@ -96,6 +96,8 @@ class StorageEditMenu(plugin_ui.SelectAndRefreshMenu, abc.ABC):
         self.render()
 
     def _on_click_delete(self):
+        if self.current_item is None:
+            return
         current_index = self.items.index(self.current_item)
         self._storage.delete(self.current_item)
         current_index = min(current_index, len(self.items) - 1)
